+++ templates/index.php
@@ -24,7 +24,7 @@
 					
             <li class="lots__item lot">
                 <div class="lot__image">
-                    <img src="<?=htmlspecialchars($lot['image']); ?>" width="350" height="260" alt="">
+                    <img src="uploads/<?=htmlspecialchars($lot['image']); ?>" width="350" height="260" alt="">
                 </div>
                 <div class="lot__info">
                     <span class="lot__category"><?=htmlspecialchars($lot['category_title']); ?></span>
